split backend_qt4agg into two parts.
One part (_backend_qt4Agg.py) has no dependencies on Gcf or pyplot,
which are all collected in backend_qt4agg.py.

Should be fully back-wards compatible.
